+++ NEWS.md
@@ -1,4 +1,8 @@
 # DPchecker 1.0.1 (under development)
+## 2025-04-24
+  * add function `test_datatable_url_attributes` to test for the appropriate attribute in the xml designation in metadata for urls. 
+  * Added unit tests for `test_datatable_url_attributes`, `test_datatable_urls_doi`, and `test_datatable_urls`
+
 ## 2025-04-16
   * remove arcticdatautils as a dependency and replace functionality with equivalent functions from EMLeditor to reduce the total number of package dependencies.